Added support for SendGrid to send to emails
This commit adds support for SendGrid to send the filepicker image links
to emails.  A simple regex match is performed on the entered phone
number.  If it matches the format (.*)@domain.xyz, it assumes it's an
email, otherwise a phone number.

You need to set SENDGRID_USER, SENDGRID_PASS, and SENDGRID_FROM_EMAIL
enviornment variables for it to work.